requirements.txt
beautifulsoup4>=4.12.0
google-api-python-client>=2.0.0
langdetect>=1.0.9
nltk>=3.8.0
numpy>=1.25.0
pandas>=2.0.0
PyYAML>=6.0.0
requests>=2.31.0
scikit-learn>=1.3.0
selenium>=4.0.0
spacy>=3.0.0
textblob>=0.17.0
tqdm>=4.0.0
youtube-transcript-api>=0.6.0
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
google-cloud-vision>=3.4.0
Pillow>=10.0.0
sentence-transformers==2.7.0
faiss-cpu>=1.12.0
